+++ src/components/openGraph/OgDescription.tsx
@@ -1,15 +1,17 @@
+import {Stack, Text} from '@sanity/ui'
 import React, {useMemo} from 'react'
 import {StringInputProps, useFormValue} from 'sanity'
-import {Stack, Text} from '@sanity/ui'
+
+import {FeedbackType} from '../../types'
 import {getOgDescriptionValidation} from '../../utils/seoUtils'
 
-const OgDescription = (props: StringInputProps) => {
+const OgDescription = (props: StringInputProps): React.ReactElement => {
   const {value, renderDefault, path} = props
 
   // Access parent object to get keywords
   const parent = useFormValue([path[0]]) as {keywords?: string[]; _type?: string}
   const isParentseoField = parent && parent?._type === 'seoFields'
-  const keywords = parent?.keywords || []
+  const keywords = useMemo(() => parent?.keywords || [], [parent?.keywords])
 
   const feedbackItems = useMemo(
     () => getOgDescriptionValidation(value || '', keywords, isParentseoField),
@@ -21,7 +23,7 @@ const OgDescription = (props: StringInputProps) => {
       {renderDefault(props)}
       {/* Validation */}
       <Stack space={2}>
-        {feedbackItems.map((item: any) => (
+        {feedbackItems.map((item: FeedbackType) => (
           <div key={item.text} style={{display: 'flex', alignItems: 'center', gap: 7}}>
             <div
               style={{