+++ src/schemas/types/index.ts
@@ -1,17 +1,19 @@
+import {SchemaTypeDefinition} from 'sanity'
+
+import {SeoFieldsPluginConfig} from '../../plugin'
 import seoFields from '..'
 import metaAttribute from './metaAttribute'
-import openGraph from './openGraph'
-import twitter from './twitter'
-import robots from './robots'
 import metaTag from './metaTag'
-import {SeoFieldsPluginConfig} from '../../plugin'
+import openGraph from './openGraph'
+import robots from './robots'
+import twitter from './twitter'
 
-export default function types(config: SeoFieldsPluginConfig = {}) {
+export default function types(config: SeoFieldsPluginConfig = {}): SchemaTypeDefinition[] {
   return [
     seoFields(config), // pass config here
     openGraph(config), // pass config here
     twitter(config), // pass config here
-    metaAttribute,
+    metaAttribute as SchemaTypeDefinition,
     metaTag,
     robots,
   ]