+++ src/utils/fieldsUtils.ts
@@ -1,94 +1,6 @@
-// import {defineField, defineType} from 'sanity'
-// import OgTitle from '../../../components/openGraph/OgTitle'
-// import OgDescription from '../../../components/openGraph/OgDescription'
+import {AllFieldKeys, SeoFieldsPluginConfig, ValidHiddenFieldKeys} from '../plugin'
 
-import {SeoFieldsPluginConfig, AllFieldKeys} from '../plugin'
-import twitter from '../schemas/types/twitter'
-
-// export default defineType({
-//   name: 'openGraph',
-//   title: 'Open Graph Settings',
-//   type: 'object',
-//   fields: [
-//     defineField({
-//       name: 'title',
-//       title: 'Open Graph Title',
-//       type: 'string',
-//       components: {
-//         input: OgTitle, // Can also wrap with a string input + preview
-//       },
-//     }),
-//     defineField({
-//       name: 'description',
-//       title: 'Open Graph Description',
-//       type: 'text',
-//       rows: 3,
-//       components: {
-//         input: OgDescription, // Can also wrap with a text area + preview
-//       },
-//     }),
-//     defineField({
-//       name: 'siteName',
-//       title: 'Open Graph Site Name',
-//       type: 'string',
-//       description: 'The name of your website. This is often the site title.',
-//     }),
-//     defineField({
-//       name: 'type',
-//       title: 'Open Graph Type',
-//       type: 'string',
-//       options: {
-//         list: [
-//           {title: 'Website', value: 'website'},
-//           {title: 'Article', value: 'article'},
-//           {title: 'Profile', value: 'profile'},
-//           {title: 'Book', value: 'book'},
-//           {title: 'Music', value: 'music'},
-//           {title: 'Video', value: 'video'},
-//           {title: 'Product', value: 'product'},
-//         ],
-//         // layout: 'radio', // Display as radio buttons
-//       },
-//       initialValue: 'website',
-//       description: 'Select the type of content for Open Graph.',
-//     }),
-
-//     // upload image or ask for url
-//     defineField({
-//       name: 'imageType',
-//       title: 'Image Type',
-//       type: 'string',
-//       options: {
-//         list: [
-//           {title: 'Upload Image', value: 'upload'},
-//           {title: 'Image URL', value: 'url'},
-//         ],
-//       },
-//       initialValue: 'upload',
-//     }),
-//     defineField({
-//       name: 'image',
-//       title: 'Open Graph Image',
-//       type: 'image',
-//       options: {
-//         hotspot: true,
-//       },
-//       hidden: ({parent}) => parent?.imageType !== 'upload',
-//       description:
-//         'Recommended size: 1200x630px (minimum 600x315px). Max file size: 5MB. Aspect ratio 1.91:1.',
-//     }),
-//     defineField({
-//       name: 'imageUrl',
-//       title: 'Open Graph Image URL',
-//       type: 'url',
-//       hidden: ({parent}) => parent?.imageType !== 'url',
-//       description:
-//         'Enter the full URL of the image. Ensure the image is accessible and meets the recommended size of 1200x630px (minimum 600x315px).',
-//     }),
-//   ],
-// })
-
-const DEFAULT_FIELD_INFO = {
+const DEFAULT_FIELD_INFO: Record<AllFieldKeys, {title: string; description: string}> = {
   title: {
     title: 'Meta Title',
     description:
@@ -159,7 +71,7 @@ const DEFAULT_FIELD_INFO = {
       'Enter the full URL of the image. Ensure the image is accessible and meets the recommended size of 1200x630px (minimum 600x315px).',
   },
   twitterCard: {
-    title: 'Card Type',
+    title: 'X Card Type',
     description: '',
   },
   twitterSite: {
@@ -198,12 +110,12 @@ const DEFAULT_FIELD_INFO = {
 export const getFieldInfo = (
   fieldName: string,
   fieldOverrides: SeoFieldsPluginConfig['fieldOverrides'],
-) => {
+): {title: string; description: string} => {
   const fieldInfo =
     (fieldOverrides && fieldOverrides[fieldName as keyof typeof fieldOverrides]) ||
     DEFAULT_FIELD_INFO[fieldName as keyof typeof DEFAULT_FIELD_INFO]
   return fieldInfo
-    ? {title: fieldInfo.title, description: fieldInfo.description}
+    ? {title: fieldInfo.title || '', description: fieldInfo.description || ''}
     : {title: '', description: ''}
 }
 
@@ -211,7 +123,7 @@ export const getFieldInfo = (
  * Check if a field should be hidden based on the current document type
  */
 export const isFieldHidden = (
-  fieldName: AllFieldKeys,
+  fieldName: ValidHiddenFieldKeys,
   config: SeoFieldsPluginConfig,
   documentType?: string,
 ): boolean => {
@@ -231,8 +143,17 @@ export const isFieldHidden = (
 /**
  * Get the hidden function for any field
  */
-export const getFieldHiddenFunction = (fieldName: AllFieldKeys, config: SeoFieldsPluginConfig) => {
-  return ({document}: {document?: any}) => {
+export const getFieldHiddenFunction = (
+  fieldName: ValidHiddenFieldKeys,
+  config: SeoFieldsPluginConfig,
+) => {
+  return ({
+    document,
+  }: {
+    document?: {
+      _type?: string
+    }
+  }): boolean => {
     const documentType = document?._type
     return isFieldHidden(fieldName, config, documentType)
   }